requirements.txt
# 1. requirements.txt
Django~=5.2.4
djangorestframework==3.14.0
drf-spectacular==0.26.2
python-decouple==3.8
django_filters
rest_framework_simplejwt
drf-yasg~=1.21.10
python-telegram-bot~=22.2
asgiref~=3.9.1
django-filter~=25.1
python-dotenv~=1.1.1